refactor: Refactor Kokeet laaja internals into private folder
ref bej-127
